+++ packages/hydrogen/src/components/LocalizationProvider/LocalizationProvider.server.tsx
@@ -3,6 +3,7 @@ import LocalizationClientProvider from './LocalizationClientProvider.client';
 import {useShopQuery} from '../../hooks/useShopQuery';
 import {LocalizationQuery} from './LocalizationQuery';
 import {Localization} from '../../graphql/graphql-constants';
+import {CacheDays} from '../../framework/CachingStrategy';
 
 /**
  * The `LocalizationProvider` component automatically queries the Storefront API's
@@ -18,7 +19,7 @@ export function LocalizationProvider({children}: {children: ReactNode}) {
     data: {localization},
   } = useShopQuery<LocalizationQuery>({
     query: Localization,
-    cache: {maxAge: 60 * 60, staleWhileRevalidate: 60 * 60 * 23},
+    cache: CacheDays(),
   });
 
   return (