+++ packages/hydrogen/src/components/Metafield/examples/metafield.example.tsx
@@ -3,5 +3,5 @@ import {Metafield} from '@shopify/hydrogen';
 export function Product({product}) {
   const metafield = product.metafields.edges.map(({node}) => node)[0];
 
-  return <Metafield metafield={metafield} />;
+  return <Metafield data={metafield} />;
 }