+++ packages/hydrogen/src/components/SelectedVariantShopPayButton/examples/selected-variant-shop-pay-button.example.tsx
@@ -2,7 +2,7 @@ import {SelectedVariantShopPayButton, ProductProvider} from '@shopify/hydrogen';
 
 export function ProductDetails({product}) {
   return (
-    <ProductProvider value={product}>
+    <ProductProvider data={product}>
       {/* Code for selecting the product variant */}
       <SelectedVariantShopPayButton />
     </ProductProvider>