+++ packages/hydrogen/src/components/index.ts
@@ -23,12 +23,10 @@ export {CartLinePrice} from './CartLinePrice';
 export {CartLineProductTitle} from './CartLineProductTitle';
 export {CartLineQuantity} from './CartLineQuantity';
 export {CartLineQuantityAdjustButton} from './CartLineQuantityAdjustButton';
-export {CartLineAttributes} from './CartLineAttributes';
 export {CartLines} from './CartLines';
 export {CartCheckoutButton} from './CartCheckoutButton';
 export {CartShopPayButton} from './CartShopPayButton';
 export {CartEstimatedCost} from './CartEstimatedCost';
-export {CartLineSelectedOptions} from './CartLineSelectedOptions';
 export {CartProvider, useCart, useInstantCheckout} from './CartProvider';
 export type {
   State,