+++ packages/playground/server-components/src/pages/index.server.jsx
@@ -9,9 +9,6 @@ export default function Index() {
   return (
     <>
       <h1>Home</h1>
-      <Link className="btn" to="/about">
-        About
-      </Link>
 
       <div className="secrets-server">
         <div>PUBLIC_VARIABLE:{import.meta.env.PUBLIC_VARIABLE || ''}|</div>
@@ -21,6 +18,10 @@ export default function Index() {
       <div className="secrets-client">
         <ClientEnv />
       </div>
+
+      <Link className="btn" to="/about">
+        About
+      </Link>
     </>
   );
 }