+++ packages/hydrogen/CHANGELOG.md
@@ -7,6 +7,57 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Removed
+
+- <CartLineSelectedOptions /> and <CartLineAttributes /> components. These components used the “function-as-a-child” pattern which doesn’t allow the `children` prop to be serialized, preventing them from being rendered within Server components.
+
+_Migration_
+
+The functionality provided by these components can be replicated using the `useCartLine()` hook instead.
+
+_Example_
+
+```tsx
+// Before
+function SomeComponent() {
+  return (
+    <>
+      <CartLineSelectedOptions as="ul" className="text-xs space-y-1">
+        {({name, value}) => (
+          <>
+            {name}: {value}
+          </>
+        )}
+      </CartLineSelectedOptions>
+      <CartLineAttributes as="ul" className="text-sm space-y-1">
+        {({key, value}) => (
+          <>
+            {key}: {value}
+          </>
+        )}
+      </CartLineAttributes>
+    </>
+  );
+}
+
+// After
+function SomeComponent() {
+  const {merchandise} = useCartLine();
+
+  return (
+    <>
+      <ul className="text-xs space-y-1">
+        {merchandise.selectedOptions.map(({name, value}) => (
+          <li key={name}>
+            {name}: {value}
+          </li>
+        ))}
+      </ul>
+    </>
+  );
+}
+```
+
 - feat: remove `fetch` workaround
 - feat: change `/react` RSC path to `/__rsc`
 - `<Model3D>` has been renamed to `<ModelViewer>`
@@ -52,6 +103,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 - The `Seo.client` component has been moved from `src/components` to `@shopify/hydrogen`. The props of the `Seo.client` component also changed to always take in `type` and `data`. Refer to the [`Seo` component reference] (../src/components/Seo/README.md) for more details. [#539](https://github.com/Shopify/hydrogen/pull/539)
 - feat: added `useUrl` hook that allows the consumer to get the current url in server or client component
 - fix: fix bug where search param is not being pass along during RSC streaming call [#623](https://github.com/Shopify/hydrogen/pull/623)
+- feat: Standardize cache control header into caching strategies by @wizardlyhel in #629
 - feat: expect collection seo by default
 - feat: Added logging option `showCacheApiStatus` and `cacheControlHeader` by @wizardlyhel in #472
 - fix: allow custom entry-client filenames
@@ -65,7 +117,23 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 - feat: pass HYDROGEN_ASSET_BASE_URL into config to set base URL for compiled assets
 - dx [breaking change]: `<Product />` and `<CartLine />` aliases have been removed; use the original components `<ProductProvider />` and `<CartLineProvider />` instead. Their nested component aliases, such as `<Product.Image />`, have also been removed; in this example you should use `<ProductImage />`.
 - feat: remove React Router on the client and introduce Hydrogen the `<Link>` component and `useNavigate` hook
+- feat [breaking change]: Merge `/src/entry-server.jsx` entry point into `App.server.jsx`
+- feat: add a default virtual entry-client in `/@shopify/hydrogen/entry-client` that can be used in `index.html`
 - fix: prevent client components from being cached during development
+- refactor: the following components had their prop name renamed. Refer to the documentation or [#627](https://github.com/Shopify/hydrogen/issues/627) for more details.
+  - `<ExternalVideo />`: renamed video prop to data
+  - `<Video />`: renamed video prop to data
+  - `<Image>`: renamed image prop to data
+  - `<MediaFile>`: renamed media prop to data
+  - `<ModelViewer>`: renamed model prop to data
+  - `<Metafield>`: renamed metafield prop to data
+  - `<Money>`: renamed money prop to data
+  - `<UnitPrice>`: renamed unitPrice prop to data, unitPriceMeasurement prop to measurement
+  - `<ProductProvider>`: renamed product prop to data
+  - `<CartProvider>`: renamed cart prop to data
+- feat [breaking change]: remove `handleEvent` in favor of `handleRequest`.
+- feat [breaking change]: remove `assetHandler` parameter in the new `handleRequest`.
+- deps: update `react-helmet-async` to 1.2.3 and remove our custome type
 
 ## 0.10.1 - 2022-01-26